% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacap.R
\name{gammacap}
\alias{gammacap}
\title{Asymptotic Covariance Matrix}
\usage{
gammacap(
  x,
  type = "mvn",
  sigmacap = NULL,
  unbiased = TRUE,
  bcap = 5000L,
  seed = NULL,
  names = TRUE,
  sep = "."
)
}
\arguments{
\item{x}{Numeric matrix, data frame, or vector.}

\item{type}{Character string.
If \code{type = "adf"},
calculate asymptotic distribution-free covariance matrix.
If \code{type = "adfnb"},
calculate nonparametric bootstrapped
asymptotic distribution-free covariance matrix.
If \code{type = "gen"},
calculate covariance matrix using the general formula.
If \code{type = "mvn"},
calculate covariance matrix with multivariate normal data.}

\item{sigmacap}{Numeric matrix.
The argument is used when \code{type = "mvn"}.
Optional argument.
Sample covariance matrix.}

\item{unbiased}{Logical.
The argument is used when \code{type = "adf"}.
If \code{unbiased = TRUE},
returns unbiased asymptotic distribution-free covariance matrix.
If \code{unbiased = FALSE},
returns consistent asymptotic distribution-free covariance matrix.}

\item{bcap}{Integer.
The argument is used when \code{type = "adfnb"}.
Number of bootstrap samples.}

\item{seed}{Integer.
The argument is used when \code{type = "adfnb"}.
Random number generation seed.}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A matrix.
}
\description{
Calculates the
covariance matrix of the unique elements
of a sample covariance matrix.
}
\examples{
set.seed(42)
n <- 1000
mu <- c(0, 0)
k <- length(mu)
sigmacap <- matrix(
  data = c(1.0, 0.5, 0.5, 1.0),
  nrow = k, ncol = k
)
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(sigmacap)
m <- matrix(data = 1, nrow = n, ncol = 1) \%*\% mu
x <- z \%*\% q + m

gammacap(x, type = "adf")
gammacap(x, type = "adfnb")
gammacap(x, type = "gen")
gammacap(x, type = "mvn")
gammacap(sigmacap = cov(x), type = "mvn")
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_adf}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_mvn}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols_hc}()},
\code{\link{gammacap_ols}()},
\code{\link{gammacapnames}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
