% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacap_adf.R
\name{gammacap_adf}
\alias{gammacap_adf}
\alias{gammacap_adfunbiased}
\title{Asymptotic Distribution-Free Covariance Matrix}
\usage{
gammacap_adf(x, unbiased = TRUE, names = TRUE, sep = ".")

gammacap_adfunbiased(gammacaptilde, sigmacaptilde, n)
}
\arguments{
\item{x}{Numeric matrix, data frame, or vector.}

\item{unbiased}{Logical.
If \code{unbiased = TRUE},
returns unbiased asymptotic distribution-free covariance matrix.
If \code{unbiased = FALSE},
returns consistent asymptotic distribution-free covariance matrix.}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}

\item{gammacaptilde}{Numeric matrix.
Consistent estimate of the asymptotic distribution-free
covariance matrix.}

\item{sigmacaptilde}{Numeric matrix.
Consistent estimate of the sample covariance matrix.}

\item{n}{Positive integer. Sample size.}
}
\value{
A matrix.
}
\description{
Calculates the asymptotic distribution-free (ADF)
covariance matrix of the unique elements
of the covariance matrix.
}
\section{Dependencies}{
\itemize{
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
set.seed(42)
n <- 1000
k <- 2
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(
  matrix(
    data = c(1.0, 0.5, 0.5, 1.0),
    nrow = k, ncol = k
  )
)
x <- z \%*\% q

gammacap_adf(x, unbiased = TRUE)
gammacap_adf(x, unbiased = FALSE)
}
\references{
Browne, M. W. (1984).
Asymptotically distribution-free methods
for the analysis of covariance structures.
British Journal of Mathematical and Statistical Psychology,
37, 62–83.
doi:\href{https://doi.org/10.1111/j.2044-8317.1984.tb00789.x}{10.1111/j.2044-8317.1984.tb00789.x}.
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_mvnadj1}()},
\code{\link{gammacap_mvnadj2}()},
\code{\link{gammacap_mvn}()},
\code{\link{gammacap_nb}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols_hc}()},
\code{\link{gammacap_ols}()},
\code{\link{gammacapnames}()},
\code{\link{gammacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
