% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacap_mvn.R
\name{gammacap_mvn}
\alias{gammacap_mvn}
\title{Asymptotic Covariance Matrix Assuming Multivariate Normal Data}
\usage{
gammacap_mvn(x, sigmacap = NULL, names = TRUE, sep = ".")
}
\arguments{
\item{x}{Numeric matrix, data frame, or vector.}

\item{sigmacap}{Numeric matrix.
Optional argument.
Sample covariance matrix.}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A matrix.
}
\description{
Calculates the covariance matrix of the unique elements
of a sample covariance matrix assuming
multivariate normal data.
}
\section{Dependencies}{
\itemize{
\item \code{\link[=dcap]{dcap()}}
\item \code{\link[=kcap]{kcap()}}
\item \code{\link[=rmvn_chol]{rmvn_chol()}} (test)
}
}

\examples{
set.seed(42)
n <- 1000
k <- 2
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(
  matrix(
    data = c(1.0, 0.5, 0.5, 1.0),
    nrow = k, ncol = k
  )
)
x <- z \%*\% q

gammacap_mvn(x)
gammacap_mvn(sigmacap = cov(x))
}
\references{
Browne, M. W., & Arminger, G. (1995).
Specification and estimation of mean-and covariance-structure  models.
Handbook of statistical modeling forthe social and behavioral sciences.
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_adf}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_nb}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols_hc}()},
\code{\link{gammacap_ols}()},
\code{\link{gammacapnames}()},
\code{\link{gammacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
