% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacap_ols.R
\name{gammacap_ols}
\alias{gammacap_ols}
\title{Asymptotic Covariance Matrix for Ordinary Least Squares Regression}
\usage{
gammacap_ols(
  x,
  type = "gen",
  adf_unbiased = TRUE,
  bcap = 5000L,
  seed = NULL,
  ke_unbiased = FALSE,
  yc = FALSE,
  names = TRUE,
  sep = "."
)
}
\arguments{
\item{x}{Object of class \code{lm}.}

\item{type}{Character string.
Type of asymptotic covariance matrix of the regressors
If \code{type = "adf"},
calculate asymptotic distribution-free covariance matrix.
If \code{type = "adfnb"},
calculate nonparametric bootstrapped
asymptotic distribution-free covariance matrix.
If \code{type = "gen"},
calculate covariance matrix using the general formula.
If \code{type = "mvn"},
calculate covariance matrix with multivariate normal data.}

\item{adf_unbiased}{Logical.
If \code{adf_unbiased = TRUE},
use unbiased asymptotic distribution-free covariance matrix.
If \code{adf_unbiased = FALSE},
use consistent asymptotic distribution-free covariance matrix.}

\item{bcap}{Integer.
Number of bootstrap samples
when \code{type = "gen"} and \code{type = "adfnb"}.}

\item{seed}{Integer.
Random number generation seed
when \code{type = "gen"} and \code{type = "adfnb"}.}

\item{ke_unbiased}{Logical.
If \code{ke_unbiased = TRUE},
use estimator
\eqn{
    \frac{
      \frac{1}{n}
      \sum_{i=1}^{n}
      \hat{\varepsilon}_{i}^{4}
    }{
      \left(
        \frac{1}{n - p - 1}
        \sum_{i=1}^{n}
        \hat{\varepsilon}_{i}^{2}
      \right)^2
    }
  }
where the denominator is the square of the unbiased estimator
of the error variance.
If \code{ke_unbiased = FALSE},
use estimator
\eqn{
    \frac{
      \frac{1}{n}
      \sum_{i=1}^{n}
      \hat{\varepsilon}_{i}^{4}
    }{
      \left(
        \frac{1}{n}
        \sum_{i=1}^{n}
        \hat{\varepsilon}_{i}^{2}
      \right)^2
    }
  }
where the denominator is the square of the consistent estimator
of the error variance.}

\item{yc}{Logical.
If \code{yc = TRUE}, order the output
following Yuan and Chan (2011) page 674.
If \code{yc = FALSE}, the order of the output
following \eqn{\mathrm{vech} \left( \Sigma \right)}.}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A matrix.
}
\description{
Calculates the asymptotic covariance matrix of the unique elements
of the covariance matrix using ordinary least squares estimates.
}
\examples{
set.seed(42)
n <- 1000
k <- 2
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(
  matrix(
    data = c(1.0, 0.5, 0.5, 1.0),
    nrow = k, ncol = k
  )
)
x <- as.data.frame(z \%*\% q)
colnames(x) <- c("y", "x")
obj <- lm(y ~ x, data = x)

gammacap_ols(obj)
}
\references{
Yuan, K.-H., & Chan, W. (2011).
Biases and standard errors of standardized regression coefficients.
Psychometrika,
76, 670–690.
doi:\href{https://doi.org/10.1007/S11336-011-9224-6}{10.1007/S11336-011-9224-6}.
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_adf}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_mvnadj1}()},
\code{\link{gammacap_mvnadj2}()},
\code{\link{gammacap_mvn}()},
\code{\link{gammacap_nb}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols_hc}()},
\code{\link{gammacapnames}()},
\code{\link{gammacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
