% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacap_ols_hc.R
\name{gammacap_ols_hc}
\alias{gammacap_ols_hc}
\title{Asymptotic Covariance Matrix for Ordinary Least Squares Regression
with Sandwich Type Adjustments}
\usage{
gammacap_ols_hc(
  x,
  type = "hc5",
  g1 = 1,
  g2 = 1.5,
  constant = 0.7,
  names = TRUE,
  sep = "."
)
}
\arguments{
\item{x}{Object.
Object of class \code{lm}.}

\item{type}{Character string.
Correction type.
Possible values are
\code{"hc0"},
\code{"hc1"},
\code{"hc2"},
\code{"hc3"},
\code{"hc4"},
\code{"hc4m"}, and
\code{"hc5"}.}

\item{g1}{Numeric.
\code{g1} value for \code{type = "hc4m"} or \code{type = "hc5"}.}

\item{g2}{Numeric.
\code{g2} value for \code{type = "hc4m"}.}

\item{constant}{Numeric.
Constant for \code{type = "hc5"}}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A matrix
}
\description{
Asymptotic Covariance Matrix for Ordinary Least Squares Regression
with Sandwich Type Adjustments
}
\examples{
set.seed(42)
n <- 1000
k <- 2
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(
  matrix(
    data = c(1.0, 0.5, 0.5, 1.0),
    nrow = k, ncol = k
  )
)
x <- as.data.frame(z \%*\% q)
colnames(x) <- c("y", "x")
obj <- lm(y ~ x, data = x)

gammacap_ols_hc(obj)
}
\references{
Dudgeon, P. (2017).
Some improvements in confidence intervals for standardized regression coefficients.
Psychometrika,
82, 928-951.
doi:\href{https://doi.org/10.1007/s11336-017-9563-z}{10.1007/s11336-017-9563-z}.
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_adf}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_mvn}()},
\code{\link{gammacap_nb}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols}()},
\code{\link{gammacapnames}()},
\code{\link{gammacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
