% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacap_ols_hc_generic.R
\name{gammacap_ols_hc_generic}
\alias{gammacap_ols_hc_generic}
\title{Asymptotic Covariance Matrix for Ordinary Least Squares Regression
with Sandwich Type Adjustments - Generic}
\usage{
gammacap_ols_hc_generic(
  x,
  h = NULL,
  type = "hc5",
  g1 = 1,
  g2 = 1.5,
  constant = 0.7,
  names = TRUE,
  sep = "."
)
}
\arguments{
\item{x}{Numeric matrix or data frame.
Data matrix for \eqn{ \{ y, x_1, \cdots, x_p \}}.}

\item{h}{Numeric vector.
Leverage values.}

\item{type}{Character string.
Type of asymptotic covariance matrix of the regressors
If \code{type = "adf"},
calculate asymptotic distribution-free covariance matrix.
If \code{type = "adfnb"},
calculate nonparametric bootstrapped
asymptotic distribution-free covariance matrix.
If \code{type = "gen"},
calculate covariance matrix using the general formula.
If \code{type = "mvn"},
calculate covariance matrix with multivariate normal data.}

\item{g1}{Numeric.
\code{g1} value for \code{type = "hc4m"} or \code{type = "hc5"}.}

\item{g2}{Numeric.
\code{g2} value for \code{type = "hc4m"}.}

\item{constant}{Numeric.
Constant for \code{type = "hc5"}}

\item{names}{Logical.
Add names.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A matrix
}
\description{
Asymptotic Covariance Matrix for Ordinary Least Squares Regression
with Sandwich Type Adjustments - Generic
}
\section{Dependencies}{
\itemize{
\item \code{\link[=gammacap_ols_hc_qcap_generic]{gammacap_ols_hc_qcap_generic()}}
}
}

\examples{
set.seed(42)
n <- 1000
k <- 2
z <- matrix(
  data = rnorm(n = n * k), nrow = n, ncol = k
)
q <- chol(
  matrix(
    data = c(1.0, 0.5, 0.5, 1.0),
    nrow = k, ncol = k
  )
)
x <- as.data.frame(z \%*\% q)
colnames(x) <- c("y", "x")
obj <- lm(y ~ x, data = x)
h <- hatvalues(obj)

gammacap_ols_hc_generic(x, h = h)
}
\references{
Dudgeon, P. (2017).
Some improvements in confidence intervals for standardized regression coefficients.
Psychometrika,
82, 928-951.
doi:\href{https://doi.org/10.1007/s11336-017-9563-z}{10.1007/s11336-017-9563-z}.
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_adf}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_mvnadj}()},
\code{\link{gammacap_mvn}()},
\code{\link{gammacap_nb}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols_hc}()},
\code{\link{gammacap_ols}()},
\code{\link{gammacapnames}()},
\code{\link{gammacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
