% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-gammacapnames.R
\name{gammacapnames}
\alias{gammacapnames}
\title{Dimension Names for the Gamma Matrix}
\usage{
gammacapnames(x, sep = ".")
}
\arguments{
\item{x}{Matrix or data frame.}

\item{sep}{Character string.
Separator for variable names.}
}
\value{
A vector.
}
\description{
Dimension Names for the Gamma Matrix
}
\examples{
x <- matrix(c(1, 0.5, 0.5, 1), nrow = 2)
colnames(x) <- rownames(x) <- c("var1", "var2")

gammacapnames(x)
}
\seealso{
Other Gamma Matrix Functions: 
\code{\link{gammacap_adfnb}()},
\code{\link{gammacap_adf}()},
\code{\link{gammacap_gen}()},
\code{\link{gammacap_mvn}()},
\code{\link{gammacap_ols_generic}()},
\code{\link{gammacap_ols_hc_generic}()},
\code{\link{gammacap_ols_hc_qcap_generic}()},
\code{\link{gammacap_ols_hc_qcap}()},
\code{\link{gammacap_ols_hc}()},
\code{\link{gammacap_ols}()},
\code{\link{gammacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Gamma Matrix Functions}
\keyword{gammaMatrix}
