% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaMatrix-package.R
\docType{package}
\name{gammaMatrix}
\alias{gammaMatrix}
\title{gammaMatrix: Asymptotic Covariance Matrix of Covariances}
\description{
A collection of functions
for generating the asymptotic covariance matrix
of covariances.
}
\references{
Browne, M. W. (1984).
Asymptotically distribution-free methods
for the analysis of covariance structures.
British Journal of Mathematical and Statistical Psychology,
37, 62–83.
doi:\href{https://doi.org/10.1111/j.2044-8317.1984.tb00789.x}{10.1111/j.2044-8317.1984.tb00789.x}.

Browne, M. W., & Arminger, G. (1995).
Specification and estimation of mean-and covariance-structure  models.
Handbook of statistical modeling forthe social and behavioral sciences.

Browne, M. W., & Arminger, G. (1995).
Asymptotically distribution-free methods
for the analysis of covariance structures.
British Journal of Mathematical and Statistical Psychology,
37, 62–83.
doi:\href{https://doi.org/10.1111/j.2044-8317.1984.tb00789.x}{10.1111/j.2044-8317.1984.tb00789.x}.

Yuan, K.-H., & Hayashi, K. (2006).
Standard errors in covariance structure models:
Asymptotics versus bootstrap.
British Journal of Mathematical and Statistical Psychology,
59, 397–417.
doi:\href{https://doi.org/10.1348/000711005X85896}{10.1348/000711005X85896}.

Yung, Y.-F., & Bentler, P. M. (1994).
Bootstrap-corrected ADF test statistics
in covariance structure analysis.
British Journal of Mathematical and Statistical Psychology,
47, 63–84.
doi:\href{https://doi.org/10.1111/j.2044-8317.1994.tb01025.x}{10.1111/j.2044-8317.1994.tb01025.x}.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{gammaMatrix}
\keyword{package}
